requirements.txt
# IMPORTANT → CPU torch (must be at top)
--extra-index-url https://download.pytorch.org/whl/cpu

# Core FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL (NeonDB)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic for settings and validation
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM — Groq API
groq

# Embeddings (CPU safe)
torch==2.2.2+cpu
transformers==4.41.2
sentence-transformers==2.7.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
rank-bm25==0.2.2
langdetect==1.0.9

# Google Drive API
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0

# Agentic Retrieval — Phase 2
langgraph==0.2.28
langchain-core==0.3.15